Add module documentation
