+++ src/citations/gather.rs
@@ -1,38 +1,8 @@
-/// Source of citations being used or defined
-///
-/// Either from bibliography file (.bib) or LaTeX file(s) (.tex)
-pub struct CitationSource<'a> {
-    pub latex_file: Option<&'a str>,
-    pub bib_file: Option<&'a str>,
-}
+//! Specify gather trait for citation
+//!
+//! The [`GatherCitations`] trait is herein defined.  Types can implement if by taking a [`CitationSource`] and returning an instance of themselves.
 
-/// Convenient implementations for construction of `CitationSource`
-impl<'a> CitationSource<'a> {
-    /// Create a new `CitationSource` object from both source files
-    pub fn new(latex_file: &'a str, bib_file: &'a str) -> Self {
-        Self {
-            latex_file: Some(latex_file),
-            bib_file: Some(bib_file),
-        }
-    }
-
-    // NOTE: the following can be uncommented if/when needed
-    // /// Create a new `CitationSource` object from just LaTeX file
-    // pub fn from_latex(latex_file: &'a str) -> Self {
-    //     Self {
-    //         latex_file: Some(latex_file),
-    //         bib_file: None,
-    //     }
-    // }
-
-    /// Create a new `CitationSource` object from just bib file
-    pub fn from_bib(bib_file: &'a str) -> Self {
-        Self {
-            latex_file: None,
-            bib_file: Some(bib_file),
-        }
-    }
-}
+use crate::source::CitationSource;
 
 /// Trait that takes a `CitationSource` and returns some collection of citations
 ///
